// src/common/enums/api-types.enum.ts
export enum ApiTypes {
  // Basic Types
  STRING = 'string',
  NUMBER = 'number',
  BOOLEAN = 'boolean',
  OBJECT = 'object',
  ARRAY = 'array',
  NULL = 'null',

  // Custom Types
  USER = 'User',
  USER_ARRAY = 'User[]',
  ACCESS_TOKEN = 'AccessToken',
  OTP_RESPONSE = 'OtpResponse',
  ERROR_RESPONSE = 'ErrorResponse',
  API_RESPONSE = 'ApiResponse',
}